Fixed bug in SBRX import, added mime codec params in -info and fixed gcc warnings

+++ src/media_tools/dash_segmenter.c
@@ -3957,9 +3957,10 @@ GF_Err gf_dash_segmenter_probe_input(GF_DashSegInput **io_dash_inputs, u32 *nb_d
                 rid = gf_dash_get_representationID(dash_inputs, *nb_dash_inputs, di->file_name, ref_track);
 
                 if (!rid) continue;
-                al_len+=strlen(rid);
-                al_len+=strlen(depID);
-                depID = gf_realloc(depID, sizeof(char)*al_len);
+                al_len += (u32) strlen(rid);
+                al_len += (u32) strlen(depID);
+
+				depID = gf_realloc(depID, sizeof(char)*al_len);
                 if (t) strcat(depID, " ");
 				strcat(depID, gf_dash_get_representationID(dash_inputs, *nb_dash_inputs, di->file_name, ref_track));
